Consolidate vendored imports into a single module
I want it to be clear what we are and aren't using. Ideally we will
eventually get rid of vendored dependencies.
