+++ conda_lock/interfaces/vendored_conda.py
@@ -1,0 +1,5 @@
+from conda_lock._vendor.conda.common.toposort import toposort
+from conda_lock._vendor.conda.models.match_spec import MatchSpec
+
+
+__all__ = ["toposort", "MatchSpec"]
