+++ conda_lock/interfaces/vendored_poetry.py
@@ -1,0 +1,32 @@
+from conda_lock._vendor.poetry.core.packages import Dependency as PoetryDependency
+from conda_lock._vendor.poetry.core.packages import Package as PoetryPackage
+from conda_lock._vendor.poetry.core.packages import (
+    ProjectPackage as PoetryProjectPackage,
+)
+from conda_lock._vendor.poetry.core.packages import URLDependency as PoetryURLDependency
+from conda_lock._vendor.poetry.core.packages import VCSDependency as PoetryVCSDependency
+from conda_lock._vendor.poetry.factory import Factory
+from conda_lock._vendor.poetry.installation.chooser import Chooser
+from conda_lock._vendor.poetry.installation.operations.uninstall import Uninstall
+from conda_lock._vendor.poetry.puzzle import Solver as PoetrySolver
+from conda_lock._vendor.poetry.repositories.pool import Pool
+from conda_lock._vendor.poetry.repositories.pypi_repository import PyPiRepository
+from conda_lock._vendor.poetry.repositories.repository import Repository
+from conda_lock._vendor.poetry.utils.env import Env
+
+
+__all__ = [
+    "Chooser",
+    "Env",
+    "Factory",
+    "PoetryDependency",
+    "PoetryPackage",
+    "PoetryProjectPackage",
+    "PoetrySolver",
+    "PoetryURLDependency",
+    "PoetryVCSDependency",
+    "Pool",
+    "PyPiRepository",
+    "Repository",
+    "Uninstall",
+]
